Resolve codec implicitly
